transatioional_webhook/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
celery>=5.3.0
redis>=5.0.0